# docs/static/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: AltMount
    url: https://github.com/javi11/altmount/issues
  description: REST API for AltMount — Usenet WebDAV server with NZB queue, health
    monitoring, ARR integration, FUSE/rclone mounting, and Stremio addon.
  license:
    name: MIT
  termsOfService: http://altmount.kipsilabs.top
  title: AltMount API
  version: "1.0"
paths:
...
  /system/browse:
    get:
      description: Lists filesystem entries at a given path for directory/file picker UIs.
      parameters:
        - description: Directory path to browse (defaults to root)
          in: query
          name: path
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Browse filesystem
      tags:
        - System
  /system/cleanup:
    post:
      description: Removes old queue items and health records based on age. Supports
        dry-run mode.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.SystemCleanupRequest"
        description: Cleanup criteria
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.SystemCleanupResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: System cleanup
      tags:
        - System
...
  /system/restart:
    post:
      description: Schedules a graceful system restart. Use force=true to skip safety
        checks.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.SystemRestartRequest"
        description: Restart options
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.SystemRestartResponse"
                    type: object
      security:
        - BearerAuth: []
      summary: Restart system
      tags:
        - System
  /system/stats:
    get:
      description: Returns combined queue, health, and system information statistics.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.SystemStatsResponse"
                    type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Get system statistics
      tags:
        - System
  /system/stats/reset:
    post:
      description: Resets accumulated system statistics counters (pool metrics,
        download counters).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Reset system statistics
      tags:
        - System
...
servers:
  - url: //localhost:8080/api
components:
...
  securitySchemes:
    ApiKeyAuth:
      in: query
      name: apikey
      type: apiKey
  schemas:
    api.APIError:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
    api.APIMeta:
      properties:
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
    api.APIResponse:
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/api.APIError"
        meta:
          $ref: "#/components/schemas/api.APIMeta"
        success:
          type: boolean
      type: object
...
    api.HealthStatsResponse:
      properties:
        checking:
          type: integer
        corrupted:
          type: integer
        healthy:
          type: integer
        pending:
          type: integer
        repair_triggered:
          type: integer
        total:
          type: integer
      type: object
...
    api.QueueStatsResponse:
      properties:
        avg_processing_time_ms:
          type: integer
        last_updated:
          type: string
        total_completed:
          type: integer
        total_failed:
          type: integer
        total_processing:
          type: integer
        total_queued:
          type: integer
      type: object
...
    api.SystemCleanupRequest:
      properties:
        dry_run:
          type: boolean
        health_older_than:
          type: string
        queue_older_than:
          type: string
      type: object
    api.SystemCleanupResponse:
      properties:
        dry_run:
          type: boolean
        health_records_removed:
          type: integer
        queue_items_removed:
          type: integer
      type: object
...
    api.SystemInfoResponse:
      properties:
        git_commit:
          type: string
        go_version:
          type: string
        start_time:
          type: string
        uptime:
          type: string
        version:
          type: string
      type: object
    api.SystemRestartRequest:
      properties:
        force:
          description: Force restart even if unsafe
          type: boolean
      type: object
    api.SystemRestartResponse:
      properties:
        message:
          type: string
        timestamp:
          type: string
      type: object
    api.SystemStatsResponse:
      properties:
        health:
          $ref: "#/components/schemas/api.HealthStatsResponse"
        queue:
          $ref: "#/components/schemas/api.QueueStatsResponse"
        system:
          $ref: "#/components/schemas/api.SystemInfoResponse"
      type: object
...